books_scrapper/requirements.txt
requests
lxml
coloredlogs
pandas
pymongo[srv]==3.12
python-dotenv